requirements.txt
"""Progressive Pinhole Ray Tracer — requirements."""

numpy>=1.24.0
PyGLM>=2.7.0
pillow>=9.0.0
pytest>=7.0.0
ruff>=0.1.0
mypy>=1.0.0
